move constants to common folder, add scipy to poetry
