+++ vision/common/constants.py
@@ -17,7 +17,8 @@ Corners: TypeAlias = NDArray[Shape["4, 2"], Float64]
 SENSOR_WIDTH: float = 13.2
 SENSOR_HEIGHT: float = 8.8
 
-# The rotation offset of the camera to the drone. The offset is applied in vision.vector_utils.pixel_intersect()
+# The rotation offset of the camera to the drone. The offset is applied
+#   in vision.vector_utils.pixel_intersect()
 # In degrees of [roll, pitch, yaw]
 # Set to [0.0, -90.0, 0.0] when the camera is facing directly downwards
 ROTATION_OFFSET: Vector = [0.0, -90.0, 0.0]
